# Faster Husimi Q with saved state