--- a/doc/apidoc/classes.rst
+++ b/doc/apidoc/classes.rst
@@ -37,6 +37,13 @@ Bloch sphere
     :members:
 
 
+Distributions
+-------------
+
+.. autoclass:: qutip.QFunc
+    :members:
+
+
 Cubic Spline
 ---------------