This is a tidied up and completed version of #934.  The first commit is credited to the original author (although I fixed up some little concerns in the PR and rebased it), because it's logically built off that PR, although in the end it's a complete reimplementation.

This is a complete rewrite of the Husimi Q calculations, to make a more formal split between the iterative (low-memory) and precomputed (faster) algorithms, giving full access to both of them.  It's a bit easier to use from a user's perspective, and it's faster than #934 as well.

The concept of the precomputed matrix is encapsulated into a class with a cached tensor, so that the same precomputed system can be used for many states of different sizes, and autoexpands (up to the memory limit) to accommodate larger states.  This is called `QFunc` as opposed to the standard function interface `qfunc`.

It also adds rather more error checking on its inputs, and adds a complete testing suite for the new implementation.

Closes #934

## Examples

Let's say we have some states, and the phase-space coordinates we want to calculate the Husimi Q function at.
```python
>>> import qutip
>>> states = [qutip.rand_dm(32, density=0.2) for _ in [None]*100]
>>> xs = np.linspace(-2, 2, 401)
```

Now the normal way of calculating the Q distribution for a single state is
```python
>>> qutip.qfunc(states[0], xs, xs)
array( ... )
```

This already has some speed advantages over the pre-PR version of `qfunc`, because it caches a lot of its intermediary results, to avoid recomputing them.  This results in calculating more matrix-vector products than before, but far fewr FLOPS overall.  By default, `qfunc` issues a warning if it exceeds a certain amount of memory, and falls back to the low-memory version.  You can control this limit with the `precompute_memory` option:
```python
>>> qutip.qfunc(states[0], xs, xs, precompute_memory=0)
qutip/qutip/wigner.py:822: UserWarning: Falling back to iterative algorithm due to lack of memory.
Needed 78.52 MB, but only allowed to use 0.00 MB.  Increase `precompute_memory` to raise limit, or
set to `None` to suppress warning.
array( ... )
```
To suppress the warning, and always use the iterative approach, you can set `precompute_memory=None`.

Now, even if you're using the `precompute_memory` option, `qfunc` still can't remember its state between calls.  This means it's inefficient.  Instead, we can use the `QFunc` class, which adds an extra step (set-up and call as opposed to just call), but it encapsulates the state.
```python
>>> qfunc_calculator = qutip.QFunc(xs, xs)
>>> husimiqs = [qfunc_calculator(state) for state in states]
```
This is much much faster.  The class version `QFunc` also takes a memory option, but since this is designed to be the precompute version, it errors out if the amount of memory needed is too great, rather than trying to fall back.  This gives users a way to put a limit on the amount of memory used:
```python
>>> qfunc_calculator_low_memory = qutip.QFunc(xs, xs, memory=0)
>>> qfunc_calculator_low_memory(states[0])
MemoryError: Refusing to precompute up to 32 basis states.  This would require 78.52 MB, but only 0 MB is allowed.
```

## Timings

These benchmarks were taken on a Macbook Pro 2.9GHz i5 (2015) with 8GB of RAM.  The different tests had different numbers of phase-space coordinates (the first number in the first column) in each dimension, and the number after 'ket' or 'dm' is the Hilbert space dimension.  The columns are:

- `original`: The state of the code before this and the previous commit
- `prev def`: Direct calls of `qutip.qfunc` with the given state and phase-space coordinates, but everything else at the default for #934
- `prev pre`: Calls of `qutip.qfunc(..., precompute=pre)`, where the calculation of `pre` was done outside the timing.  This times how long it would take to call in a loop many times.
- `cur def`: Direct calls of `qutip.qfunc` with the given state and phase-space coordinates, but everything else at the default for this commit.
- `cur pre`: Calls to a pre-constructed `qutip.QFunc` instance (which was excluded from the timings) for this commit.  This times how long it would take to call in a loop many times.

```text
            │ original  │ prev def  │ prev pre  │  cur def  │  cur pre
────────────┼───────────┼───────────┼───────────┼───────────┼──────────
11,  ket 5  │ 126.07 µs │ 133.02 µs │  89.63 µs │ 108.95 µs │  25.12 µs
11,  ket 32 │ 192.49 µs │ 197.88 µs │  90.91 µs │ 172.13 µs │  29.33 µs
11,  dm 5   │ 481.68 µs │ 440.45 µs │ 316.93 µs │ 304.86 µs │ 161.00 µs
11,  dm 32  │   4.69 ms │   1.87 ms │   1.66 ms │   1.07 ms │ 823.49 µs
            │           │           │           │           │
101, ket 5  │ 570.59 µs │ 566.65 µs │ 501.86 µs │ 463.96 µs │ 329.25 µs
101, ket 32 │   1.61 ms │   1.43 ms │ 751.97 µs │   1.34 ms │ 507.20 µs
101, dm 5   │   2.51 ms │   3.51 ms │   2.20 ms │   2.62 ms │   1.72 ms
101, dm 32  │  42.91 ms │  26.84 ms │  19.73 ms │  23.30 ms │  16.30 ms
            │           │           │           │           │
401, ket 5  │  10.56 ms │  13.50 ms │   7.91 ms │  11.95 ms │   5.28 ms
401, ket 32 │  34.44 ms │  43.73 ms │  10.37 ms │  41.17 ms │   7.70 ms
401, dm 5   │  46.85 ms │  57.51 ms │  33.19 ms │  52.30 ms │  26.71 ms
401, dm 32  │   1.05  s │ 411.32 ms │ 273.93 ms │ 428.32 ms │ 246.59 ms
```

You can see that there are huge benefits to the precomputation, even compared to the previous precomputation implementation.  This makes calling in a loop much much more efficient.


#### Changelog

Large feature, authored by @dweigand (Daniel Weigand) and me (Jake).

Add new Husimi Q algorithms, improving the speed for density matrices, and giving a near order-of-magnitude improvement when calculating the Q function for many different states (using the new `qutip.QFunc` class, instead of the `qutip.qfunc` function).