@jakelishman 2021-06-20
(Also apparently I tidied a couple of imports and docstrings in `wigner.py` as well, which I forgot...)