@jakelishman 2021-07-01
@Ericgig: yeah, it's not a perfect naming choice.  Can you think of a better one?  It ended up like that because I started off just calling in `_qfunc_alpha_matrix`, then realised about halfway through that I was actually constructing the conjugate of the matrix formed by stacking coherent states (since that's how it's used), but the name isn't really ideal.

@quantshah: I'm just thinking now you might want to have a look at this - you had some thoughts about user interfaces for class-based creation schemes last year, with regards to the displacement operator, which I'll build up into a proper PR at some point as well.  Don't worry if you've no time, though.