@quantshah 2021-07-01
Yes I will have a look this weekend.

On Thu, 1 Jul 2021 at 14:32, Jake Lishman ***@***.***> wrote:

> @Ericgig <https://github.com/Ericgig>: yeah, it's not a perfect naming
> choice. Can you think of a better one? It ended up like that because I
> started off just calling in _qfunc_alpha_matrix, then realised about
> halfway through that I was actually constructing the conjugate of the
> matrix formed by stacking coherent states (since that's how it's used), but
> the name isn't really ideal.
>
> @quantshah <https://github.com/quantshah>: I'm just thinking now you
> might want to have a look at this - you had some thoughts about user
> interfaces for class-based creation schemes last year, with regards to the
> displacement operator, which I'll build up into a proper PR at some point
> as well. Don't worry if you've no time, though.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/qutip/qutip/pull/1583#issuecomment-872207526>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABVFIBH3GWFKFVFVL4ZBR2DTVRN6VANCNFSM47AL5RKA>
> .
>