@Ericgig 2021-07-01
I am not great with names myself, but I would go with `grid`, or `mesh`.