@jakelishman 2021-07-09
Oh, also to answer the question more explicitly about `_qfunc_alpha_matrix.__call__`: _that_ function always re-computes the matrix of alphas, returning the relevant bits of the Fock space.  The caching behaviour is in the `QFunc` class itself, which manages how many Fock states it needs (see `QFunc._alphas`, which tries to pull the matrix from a cache, or expands the cache if necessary).  `QFunc.__call__` calculates the Q function of any quantum state over the given phase-space coordinates, no matter how large the size of Hilbert space of the state is - that means it's got feature parity with `qfunc` in the sense that both
```python
iterative = functools.partial(qutip.qfunc, xvec=xs, yvec=ys)
classbased = qutip.QFunc(xs, ys)
```
can be called with the same objects, and will return the same result, even if the objects change size in between.  I see this as sensible functionality in general - if you're plotting a few graphs, it's quite possible that you haven't calculated all your states with the same amount of Fock-space truncation (I frequently scaled the calculation sizes in my PhD for speed), so it's not very useful if the class-based version can only handle fixed Hilbert spaces.