@quantshah 2021-07-09
Thanks Jake, I understand better how the cache is working. Okay let me suggest a few comments within the code so that we (or at least I) can keep track of how this implementation works. If you think they are sensible then feel free to accept and then we can merge this.